django/home_service/requirements.txt
Django
djangorestframework
django-allauth
python-dotenv
django-cors-headers
djangorestframework-simplejwt
psycopg2-binary
Pillow